app/requirements.txt
alembic==1.6.5
boto3==1.21.16
botocore==1.24.17
certifi==2024.8.30
click==7.1.2
distlib==0.3.8
filelock==3.16.1
Flask==2.0.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==2.0.1
packaging==24.1
pipenv==2024.0.3
platformdirs==4.3.6
psycopg2-binary==2.9.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
s3transfer==0.5.2
six==1.15.0
SQLAlchemy==1.3.24
urllib3==1.26.8
virtualenv==20.26.5
Werkzeug==2.0.1
WTForms==2.3.3
